+++ client/js/domain/BatteryStatusTable.js
@@ -1,159 +1,24 @@
 import React from 'react'
-import { Row } from 'react-bootstrap'
-import { AppContext } from '../context/AppContext'
+import ObjectTable, { SELECTION, DATA_METHOD } from './ObjectTable'
+import config from '../config'
+//import EditPatientForm from './EditPatientForm'
 import { batteryTableColumn } from '../config/tables'
-import { toast } from 'react-toastify'
-import { setSuccessMessage, setErrorMessage } from '../helper/messageGenerator'
-import API from '../api'
-import Table from '../components/Table'
-import ObjectFilterBar from '../components/ObjectFilterBar'
-import { SET_TABLE_SELECTION } from '../reducer/action'
-import { keyBy } from 'lodash'
+//import BatteryStatusTable from './BatteryStatusTable'
+import EditPatientForm from './EditPatientForm'
+import { PERSON } from '../config/wordMap'
 
-class BatteryStatusTable extends React.Component {
-	static contextType = AppContext
-
-	state = {
-		data: [],
-		tabIndex: 0,
-		locale: this.context.locale.abbr,
-		objectFilter: [],
-		filteredData: [],
-		filterSelection: [],
-		areaTable: [],
-	}
-
-	componentDidUpdate = (prevProps, prevState) => {
-		const { locale } = this.context
-		if (locale.abbr !== prevState.locale) {
-			this.getTrackingData()
-		}
-	}
-
-	componentDidMount = () => {
-		this.getTrackingData()
-	}
-
-	getTrackingData = async (callback) => {
-		const { locale, stateReducer } = this.context
-		const [{ area }] = stateReducer
-		const { auth } = this.context
-		const { user } = auth
-		const BatteryDataPromise = API.Tracking.getTrackingData({
-			areaIds: [area.id],
-			locale: locale.addr,
-		})
-
-		const areaTablePromise = API.Area.getAreaTableByUserId({
-			userId: user.id,
-		})
-
-		const [BatteryDataRes, areaTableRes] = await Promise.all([
-			BatteryDataPromise,
-			areaTablePromise,
-		])
-
-		if (BatteryDataRes && areaTableRes) {
-			this.setMessage('clear')
-			const areaDataMap = keyBy(areaTableRes.data, 'id')
-			const data = BatteryDataRes.data.map((item) => {
-				item.mac_address = item.mac_address
-					? item.mac_address
-					: locale.texts.NON_BINDING
-				item.area_name = {
-					value: areaDataMap[item.area_id],
-					label:
-						areaDataMap[item.area_id] &&
-						areaDataMap[item.area_id].readable_name,
-					id: item.area_id,
-				}
-				return item
-			})
-
-			const areaSelection = areaTableRes.data.map((area) => {
-				return {
-					value: area.name,
-					label: area.readable_name,
-				}
-			})
-			this.setState(
-				{
-					data,
-					filteredData: data,
-					areaTable: areaTableRes.data,
-					areaSelection,
-					locale: locale.abbr,
-				},
-				callback
-			)
-			this.clearSelection()
-		} else {
-			this.setMessage('error', 'connect to database failed', true)
-		}
-	}
-	clearSelection = () => {
-		const [, dispatch] = this.context.stateReducer
-		dispatch({
-			type: SET_TABLE_SELECTION,
-			value: [],
-		})
-	}
-	setMessage = (type, msg, isSetting) => {
-		switch (type) {
-			case 'success':
-				this.toastId = setSuccessMessage(msg)
-				break
-			case 'error':
-				if (isSetting && !this.toastId) {
-					this.toastId = setErrorMessage(msg)
-				}
-				break
-			case 'clear':
-				this.toastId = null
-				toast.dismiss(this.toastId)
-				break
-		}
-	}
-
-	render() {
-		const { locale } = this.context
-		const filteredAttribute = ['name', 'type', 'area', 'macAddress', 'acn']
-		return (
-			<>
-				<Row>
-					<ObjectFilterBar
-						onFilterUpdated={({ objectFilter, filteredData }) => {
-							this.setState({
-								objectFilter,
-								filteredData,
-							})
-							this.clearSelection()
-						}}
-						oldObjectFilter={this.state.objectFilter}
-						objectList={this.state.data}
-						selectionList={[
-							{
-								label: locale.texts.SEARCH,
-								attribute: filteredAttribute,
-								source: 'search bar',
-							},
-							{
-								label: locale.texts.AREA,
-								attribute: ['area'],
-								options: this.state.areaSelection,
-								source: 'area select',
-							},
-						]}
-					/>
-				</Row>
-				<Table
-					style={{ maxHeight: '85vh' }}
-					data={this.state.filteredData}
-					columns={batteryTableColumn}
-				/>
-			</>
-		)
-	}
+const BatteryStatusTable=()=>{
+    return (
+        <ObjectTable
+            objectTypes={[config.OBJECT_TYPE.PERSON, config.OBJECT_TYPE.DEVICE]}
+            filteredAttribute={['name', 'type', 'area', 'macAddress', 'acn']}
+            enabledSelection={[SELECTION.AREA]}
+            columns={batteryTableColumn}
+            EditedForm={EditPatientForm}
+            objectApiMode={PERSON}
+            dataMethod={DATA_METHOD.TRACKING}
+        />
+    )
 }
 
 export default BatteryStatusTable
